requirements.txt
Django>=3.0,<=4.1
django-environ==0.9
djangorestframework<=3.13.1
psycopg2-binary==2.9.3
